fix(reset): align smart-reset config and async hook behavior
